Refactor and reorganize function parser structure.
Moved `_FunctionParser` to a new `base_parser` module for improved modularity and maintainability. Removed unused tests and adjusted imports accordingly. This change enhances code organization and simplifies future updates.
